+++ config/laravel-model-caching.php
@@ -3,4 +3,5 @@
 return [
     'store' => env('MODEL_CACHE_STORE'),
     'debug' => env('MODEL_CACHE_DEBUG', false),
+    'cache-prefix' => '',
 ];