+++ src/CachedBuilder.php
@@ -21,14 +21,11 @@ class CachedBuilder extends EloquentBuilder
             return parent::avg($column);
         }
 
-        return $this->cache($this->makeCacheTags())
-            ->remember(
-                $this->makeCacheKey(['*'], null, "-avg_{$column}"),
-                $this->cacheTime,
-                function () use ($column) {
-                    return parent::avg($column);
-                }
-            );
+        $arguments = func_get_args();
+        $cacheKey = $this->makeCacheKey(['*'], null, "-avg_{$column}");
+        $method = 'avg';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
     public function count($columns = ['*'])
@@ -37,14 +34,11 @@ class CachedBuilder extends EloquentBuilder
             return parent::count($columns);
         }
 
-        return $this->cache($this->makeCacheTags())
-            ->remember(
-                $this->makeCacheKey(['*'], null, "-count"),
-                $this->cacheTime,
-                function () use ($columns) {
-                    return parent::count($columns);
-                }
-            );
+        $arguments = func_get_args();
+        $cacheKey = $this->makeCacheKey(['*'], null, "-count");
+        $method = 'count';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
     public function cursor()
@@ -53,14 +47,11 @@ class CachedBuilder extends EloquentBuilder
             return collect(parent::cursor());
         }
 
-        return $this->cache($this->makeCacheTags())
-            ->remember(
-                $this->makeCacheKey(['*'], null, "-cursor"),
-                $this->cacheTime,
-                function () {
-                    return collect(parent::cursor());
-                }
-            );
+        $arguments = func_get_args();
+        $cacheKey = $this->makeCacheKey(['*'], null, "-cursor");
+        $method = 'cursor';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
     public function delete()
@@ -80,14 +71,11 @@ class CachedBuilder extends EloquentBuilder
             return parent::find($id, $columns);
         }
 
-        return $this->cache($this->makeCacheTags())
-            ->remember(
-                $this->makeCacheKey($columns, $id),
-                $this->cacheTime,
-                function () use ($id, $columns) {
-                    return parent::find($id, $columns);
-                }
-            );
+        $arguments = func_get_args();
+        $cacheKey = $this->makeCacheKey($columns);
+        $method = 'find';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
     public function first($columns = ['*'])
@@ -96,14 +84,11 @@ class CachedBuilder extends EloquentBuilder
             return parent::first($columns);
         }
 
-        return $this->cache($this->makeCacheTags())
-            ->remember(
-                $this->makeCacheKey($columns, null, '-first'),
-                $this->cacheTime,
-                function () use ($columns) {
-                    return parent::first($columns);
-                }
-            );
+        $arguments = func_get_args();
+        $cacheKey = $this->makeCacheKey($columns);
+        $method = 'first';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
     public function get($columns = ['*'])
@@ -112,14 +97,11 @@ class CachedBuilder extends EloquentBuilder
             return parent::get($columns);
         }
 
-        return $this->cache($this->makeCacheTags())
-            ->remember(
-                $this->makeCacheKey($columns),
-                $this->cacheTime,
-                function () use ($columns) {
-                    return parent::get($columns);
-                }
-            );
+        $arguments = func_get_args();
+        $cacheKey = $this->makeCacheKey($columns);
+        $method = 'get';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
     public function max($column)
@@ -128,14 +110,11 @@ class CachedBuilder extends EloquentBuilder
             return parent::max($column);
         }
 
-        return $this->cache($this->makeCacheTags())
-            ->remember(
-                $this->makeCacheKey(['*'], null, "-max_{$column}"),
-                $this->cacheTime,
-                function () use ($column) {
-                    return parent::max($column);
-                }
-            );
+        $arguments = func_get_args();
+        $cacheKey = $this->makeCacheKey(['*'], null, "-max_{$column}");
+        $method = 'max';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
     public function min($column)
@@ -144,14 +123,11 @@ class CachedBuilder extends EloquentBuilder
             return parent::min($column);
         }
 
-        return $this->cache($this->makeCacheTags())
-            ->remember(
-                $this->makeCacheKey(['*'], null, "-min_{$column}"),
-                $this->cacheTime,
-                function () use ($column) {
-                    return parent::min($column);
-                }
-            );
+        $arguments = func_get_args();
+        $cacheKey = $this->makeCacheKey(['*'], null, "-min_{$column}");
+        $method = 'min';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
     public function pluck($column, $key = null)
@@ -161,12 +137,11 @@ class CachedBuilder extends EloquentBuilder
         }
 
         $keyDifferentiator = "-pluck_{$column}" . ($key ? "_{$key}" : "");
+        $arguments = func_get_args();
         $cacheKey = $this->makeCacheKey([$column], null, $keyDifferentiator);
+        $method = 'pluck';
 
-        return $this->cache($this->makeCacheTags())
-            ->remember($cacheKey, $this->cacheTime, function () use ($column, $key) {
-                return parent::pluck($column, $key);
-            });
+        return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
     public function sum($column)
@@ -175,14 +150,11 @@ class CachedBuilder extends EloquentBuilder
             return parent::sum($column);
         }
 
-        return $this->cache($this->makeCacheTags())
-            ->remember(
-                $this->makeCacheKey(['*'], null, "-sum_{$column}"),
-                $this->cacheTime,
-                function () use ($column) {
-                    return parent::sum($column);
-                }
-            );
+        $arguments = func_get_args();
+        $cacheKey = $this->makeCacheKey(['*'], null, "-sum_{$column}");
+        $method = 'sum';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
     public function value($column)
@@ -191,12 +163,57 @@ class CachedBuilder extends EloquentBuilder
             return parent::value($column);
         }
 
-        return $this->cache($this->makeCacheTags())
+        $arguments = func_get_args();
+        $cacheKey = $this->makeCacheKey(['*'], null, "-value_{$column}");
+        $method = 'value';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
+    }
+
+    public function cachedValue(array $arguments, string $cacheKey, string $method)
+    {
+        $cacheTags = $this->makeCacheTags();
+        $hashedCacheKey = sha1($cacheKey);
+
+        $result = $this->retrieveCachedValue(
+            $arguments,
+            $cacheKey,
+            $cacheTags,
+            $hashedCacheKey,
+            $method
+        );
+
+        if ($result['key'] !== $cacheKey) {
+            cache()->tags($cacheTags)->forget($hashedCacheKey);
+        }
+
+        $result = $this->retrieveCachedValue(
+            $arguments,
+            $cacheKey,
+            $cacheTags,
+            $hashedCacheKey,
+            $method
+        );
+
+        return $result['value'];
+    }
+
+    protected function retrieveCachedValue(
+        array $arguments,
+        string $cacheKey,
+        array $cacheTags,
+        string $hashedCacheKey,
+        string $method
+    ) {
+        return $this->cache($cacheTags)
             ->remember(
-                $this->makeCacheKey(['*'], null, "-value_{$column}"),
+                $hashedCacheKey,
                 $this->cacheTime,
-                function () use ($column) {
-                    return parent::value($column);
+                function () use ($arguments, $cacheKey, $method) {
+                    return [
+                        'key' => $cacheKey,
+                        'value' => parent::{$method}(...$arguments),
+                    ];
                 }
             );
     }