+++ CHANGELOG.md
@@ -4,7 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.2.27] - 16 Feb 2018
+## [0.2.31] - 18 Feb 2018
+### Added
+- optional cache key prefixing.
+
+## [0.2.30] - 18 Feb 2018
+### Changed
+- detection of Cachable trait to use `class_uses()` instead of looking for
+  method.
+
+## [0.2.29] - 18 Feb 2018
+### Added
+- hash collision detection and prevetion.
+
+## [0.2.28] - 18 Feb 2018
+### Changed
+- disabling of cache from using session to use cache-key instead.
+
+## [0.2.27] - 17 Feb 2018
 ### Fixed
 - the erroneous use of `arrayEmpty()` function, changed to simple `count()`.